requirements.txt
fastapi==0.110.1
pydantic==1.10.21
uvicorn==0.23.2
numpy==1.26.4
Pillow==10.3.0
reportlab==4.0.8
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.0
typing-extensions==4.12.2
